Implemented the StoreBælt system. It handles creating the vehicle, so the user of the library doesn't have to know the other libray.
However using a string to check the type feels to me like a weak link. The rework required in the next task might be an opportunity to fix that.

Tests failed initially because I first forgot that today was a weekend day and I lazily used DateTime.Now and then because I forgot to implement the discount being only for cars while checking the MC-price (Which I only did to bring up code coverage).

Code coverage is alright when running for all tests. The StoreBælt tests does not check for things already checked in the other library I hope that is alright.

All tests succeeded.